+++ CHANGES.md
@@ -10,6 +10,7 @@
 * add 'can_vote' to admin user.
 * refactor to PSR 4
 * add hooks to WP-Polls to remove names from casted ballots
+* add memo field to EDD emails
 
 #### 1.9.0
 * remove a few more admin menu items